#2 名前付プレースホルダーへ統一いたしました。

+++ lib/model/cart.php
@@ -14,11 +14,11 @@
 function cart_is_exists_item($db, $user_id, $item_id) {
 	$sql = <<<EOD
 SELECT item_id, amount FROM carts
-WHERE user_id = ? AND item_id = ?
+WHERE user_id = :user_id AND item_id = :item_id
 EOD;
     $params = array(
-        $user_id,
-        $item_id
+        ':user_id' => $user_id,
+        ':item_id' => $item_id
     );
     $cart = db_select($sql, $db, $params);
 	return empty($cart) === false;
@@ -34,10 +34,10 @@ function cart_total_price($db, $user_id) {
 SELECT sum(price * amount) as total_price
  FROM carts JOIN items
  ON carts.item_id = items.id
- WHERE items.status = 1 AND user_id = ?
+ WHERE items.status = 1 AND user_id = :user_id
 EOD;
 	$params = array(
-	    $user_id
+	    ':user_id' => $user_id
 	);
 	$row = db_select_one($sql, $db, $params);
 	if (empty($row)) {
@@ -56,10 +56,10 @@ function cart_list($db, $user_id) {
  SELECT carts.id, item_id, name, price, img, amount, (price * amount) as amount_price
  FROM carts JOIN items
  ON carts.item_id = items.id
- WHERE items.status = 1 AND user_id = ?
+ WHERE items.status = 1 AND user_id = :user_id
 EOD;
 	$params = array(
-	    $user_id
+	    ':user_id' => $user_id
 	);
 	return db_select($sql, $db, $params);
 }
@@ -77,20 +77,20 @@ function cart_regist($db, $user_id, $item_id) {
 		$sql = <<<EOD
 UPDATE carts
  SET amount = amount + 1 , update_date = NOW()
- WHERE user_id =  ? AND item_id = ?
+ WHERE user_id = :user_id AND item_id = :item_id
 EOD;
 		$params = array(
-		    $user_id,
-		    $item_id
+		    ':user_id' => $user_id,
+		    ':item_id' => $item_id
 		);
 	} else {
 		$sql = <<<EOD
 INSERT INTO carts (user_id, item_id, amount, create_date, update_date)
-VALUES (?, ?}, 1, NOW(), NOW())
+VALUES (:user_id, :item_id, 1, NOW(), NOW())
 EOD;
     	$params = array(
-    	    $user_id,
-    	    $item_id
+    	    ':user_id' => $user_id,
+    	    ':item_id' => $item_id
     	);
 	}
 	return db_update($db, $sql, $params);
@@ -106,13 +106,13 @@ EOD;
 function cart_update($db, $id, $user_id, $amount) {
 	$sql = <<<EOD
 UPDATE carts
- SET amount = ?, update_date = NOW()
- WHERE id = ? AND user_id = ?
+ SET amount = :amount, update_date = NOW()
+ WHERE id = :id AND user_id = :user_id
 EOD;
 	$params = array(
-	    $amount,
-	    $id,
-	    $user_id
+	    ':amount' => $amount,
+	    ':id' => $id,
+	    ':user_id' => $user_id
 	);
 	return db_update($db, $sql, $params);
 }
@@ -126,11 +126,11 @@ EOD;
 function cart_delete($db, $id, $user_id) {
 	$sql = <<<EOD
 DELETE FROM carts
- WHERE id = ? AND user_id = ?
+ WHERE id = :id AND user_id = :user_id
 EOD;
 	$params = array(
-	    $id,
-	    $user_id
+	    ':id' => $id,
+	    ':user_id' => $user_id
 	);
 	return db_update($db, $sql, $params);
 }
@@ -141,9 +141,9 @@ EOD;
  * @return int
  */
 function cart_clear($db, $user_id) {
-	$sql = 'DELETE FROM carts WHERE user_id = ?';
+	$sql = 'DELETE FROM carts WHERE user_id = :user_id';
 	$params = array(
-	    $user_id
+	    ':user_id' => $user_id
 	);
 	return db_update($db, $sql, $params);
 }
